pyproject.toml
[project]
name = "lyricus"
version = "0.1.0"
description = "LLM による音楽生成フレームワーク"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-ollama>=1.1.0",
    "loguru>=0.7.3",
    "music21>=9.9.1",
    "pydantic>=2.13.2",
    "pydantic-settings>=2.13.1",
    "pyfluidsynth>=1.3.4",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "tqdm>=4.67.3",
]

[project.scripts]
lyricus = "lyricus.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lyricus"]